+++ packages/components/src/index.ts
@@ -4,3 +4,4 @@ export * from './components/radio/index.js';
 export * from './components/toggle/index.js';
 export * from './components/input/index.js';
 export * from './components/select/index.js';
+export * from './components/loader/index.js';